code cleanup, merged scroll functionality into a util, removed dedicated scrollDown button from hero section
